+++ components/SupportPrograms/SupportProgramFilters/FilterDetail/AreaFilterDetail.tsx
@@ -1,7 +1,7 @@
 import { Area } from '../SupportProgramFilters.types';
 import { useClientFilter, useFilterByQueryString } from '../SupportProgramFilters.hooks';
 import { identity } from '../../SupportPrograms.utils';
-import FilterDetail from './FilterDetailRefactor';
+import FilterDetailModal from './FilterDetailModal';
 import FilterItem from '../FilterItem';
 
 type Props = {
@@ -25,7 +25,7 @@ function AreaFilterDetail({ title, list }: Props) {
   };
 
   return (
-    <FilterDetail
+    <FilterDetailModal
       resetItems={toggleState('all')}
       title={title}
       onApply={onApply}
@@ -38,7 +38,7 @@ function AreaFilterDetail({ title, list }: Props) {
           </FilterItem>
         </li>
       ))}
-    </FilterDetail>
+    </FilterDetailModal>
   );
 }
